fix: renamed files and folders for readability
